Rename VViewSwitcherNode -> VViewSwitcherEntry
