+++ InteractionBase/styles/item/VViewSwitcherEntry/default
